add first Map entities

+++ src/main/java/com/ustudents/examples/scenes/EcsExample2.java
@@ -13,6 +13,7 @@ import com.ustudents.farmland.scene.MainMenu;
 import imgui.ImGui;
 import imgui.flag.ImGuiCond;
 import org.joml.Vector2f;
+import org.joml.Vector4f;
 import org.joml.Vector4i;
 
 public class EcsExample2 extends Scene {
@@ -46,7 +47,7 @@ public class EcsExample2 extends Scene {
                 int textureRegionX = 24 * random.generateInRange(1, 5);
                 int textureRegionY = 24 * random.generateInRange(1, 5);
                 grass.addComponent(SpriteComponent.class, texture,
-                        new Vector4i(textureRegionX, textureRegionY, 24, 24));
+                        new Vector4f(textureRegionX, textureRegionY, 24, 24));
             }
         }
     }
